Fix BossBarPacketListener.java

+++ src/protocolsupportpocketstuff/hacks/holograms/HologramsPacketListener.java
@@ -187,7 +187,7 @@ public class HologramsPacketListener extends Connection.PacketListener {
 			}
 
 			if (metaType == 0) {
-				long peBaseFlags = (Integer) dw.getValue();
+				long peBaseFlags = (Long) dw.getValue();
 
 				isInvisible = ((peBaseFlags & (1 << (7 - 1))) != 0);
 				showNametag = ((peBaseFlags & (1 << (15 - 1))) != 0);
